Fixing MinGW warnings in rabema.
git-svn-id: http://svn.seqan.de/seqan/trunk@8125 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/rabema/evaluation.h
@@ -110,7 +110,7 @@ int computeQualityAlignmentScore(TAlign const & align,
                 int x = 0;
                 x += getQualityValue(read[toSourcePosition(rowRead, i) - 1]);
                 x += getQualityValue(read[toSourcePosition(rowRead, i)]);
-                result -= ceil(1.0 * x / 2);
+                result -= static_cast<int>(ceil(1.0 * x / 2));
             }
         } else {
             result += score(scoreMatrix, readChar, contigChar) * getQualityValue(read[toSourcePosition(rowRead, i)]);
